+++ src/client.ts
@@ -806,6 +806,7 @@ import {
   SiprecConnectorUpdateResponse,
   SiprecConnectors,
 } from './resources/siprec-connectors';
+import { SpeechToText, SpeechToTextTranscribeParams } from './resources/speech-to-text';
 import {
   SubNumberOrder,
   SubNumberOrderCancelResponse,
@@ -853,6 +854,14 @@ import {
   TexmlApplications,
   TexmlApplicationsDefaultFlatPagination,
 } from './resources/texml-applications';
+import {
+  TextToSpeech,
+  TextToSpeechGenerateParams,
+  TextToSpeechGenerateResponse,
+  TextToSpeechListVoicesParams,
+  TextToSpeechListVoicesResponse,
+  TextToSpeechStreamParams,
+} from './resources/text-to-speech';
 import {
   UsageReportGetOptionsParams,
   UsageReportGetOptionsResponse,
@@ -1464,15 +1473,6 @@ import {
 } from './resources/sim-cards/sim-cards';
 import { Storage, StorageListMigrationSourceCoverageResponse } from './resources/storage/storage';
 import { Texml, TexmlSecretsParams, TexmlSecretsResponse } from './resources/texml/texml';
-import {
-  StreamClientEvent,
-  StreamServerEvent,
-  TextToSpeech,
-  TextToSpeechGenerateParams,
-  TextToSpeechGenerateResponse,
-  TextToSpeechListVoicesParams,
-  TextToSpeechListVoicesResponse,
-} from './resources/text-to-speech/text-to-speech';
 import {
   CreateVerificationResponse,
   Verification,
@@ -2856,6 +2856,10 @@ export class Telnyx {
    */
   voiceDesigns: API.VoiceDesigns = new API.VoiceDesigns(this);
   x402: API.X402 = new API.X402(this);
+  /**
+   * Speech to text command operations
+   */
+  speechToText: API.SpeechToText = new API.SpeechToText(this);
 }
 
 Telnyx.Legacy = Legacy;
@@ -3016,6 +3020,7 @@ Telnyx.WhatsappMessageTemplates = WhatsappMessageTemplates;
 Telnyx.VoiceClones = VoiceClones;
 Telnyx.VoiceDesigns = VoiceDesigns;
 Telnyx.X402 = X402;
+Telnyx.SpeechToText = SpeechToText;
 
 export declare namespace Telnyx {
   export type RequestOptions = Opts.RequestOptions;
@@ -4459,10 +4464,9 @@ export declare namespace Telnyx {
     TextToSpeech as TextToSpeech,
     type TextToSpeechGenerateResponse as TextToSpeechGenerateResponse,
     type TextToSpeechListVoicesResponse as TextToSpeechListVoicesResponse,
-    type StreamClientEvent as StreamClientEvent,
-    type StreamServerEvent as StreamServerEvent,
     type TextToSpeechGenerateParams as TextToSpeechGenerateParams,
     type TextToSpeechListVoicesParams as TextToSpeechListVoicesParams,
+    type TextToSpeechStreamParams as TextToSpeechStreamParams,
   };
 
   export {
@@ -4715,6 +4719,8 @@ export declare namespace Telnyx {
   };
 
   export { X402 as X402 };
+
+  export { SpeechToText as SpeechToText, type SpeechToTextTranscribeParams as SpeechToTextTranscribeParams };
 
   export type APIError = API.APIError;
   export type AvailablePhoneNumbersMetadata = API.AvailablePhoneNumbersMetadata;