+++ src/resources/text-to-speech/index.ts
@@ -8,5 +8,6 @@ export {
   type StreamServerEvent,
   type TextToSpeechGenerateParams,
   type TextToSpeechListVoicesParams,
+  type TextToSpeechStreamParams,
 } from './text-to-speech';
 export { TextToSpeechWS } from './ws';